Further refactor of image exporting code
All of the raster image out functions in export_task.py have been condensed down into one to_raster() function.

- Improved an error printout in orxport.py.

+++ export_task.py
@@ -29,101 +29,37 @@ def to_svg(emoji_svg, out_path, license=None):
 
 
 
-
-def to_png(emoji_svg, out_path, renderer, size, name):
-    """
-    PNG Exporting function. Creates temporary SVGs
-    first before converting to PNG.
-    """
-    # saving SVG to a temporary file
-    tmp_name = '.tmp' + name + '.svg'
-
-    # try to write a temporary SVG.
-    image_proc.write_temp_svg(emoji_svg, tmp_name)
-    # try to render the SVG.
-    image_proc.render_svg(tmp_name, out_path, renderer, size)
-
-    # delete temporary files
-    os.remove(tmp_name)
-
-
-
-
-
-
-
-def to_webp(emoji_svg, out_path, renderer, size, name):
-    """
-    WebP Exporting function. Creates temporary SVGs and PNGs
-    first before converting to WebP.
-    """
+def to_raster(emoji_svg, out_path, renderer, format, size, name):
 
     tmp_svg_path = '.tmp' + name + '.svg'
     tmp_png_path = '.tmp' + name + '.png'
 
     # try to write a temporary SVG.
     image_proc.write_temp_svg(emoji_svg, tmp_svg_path)
-    # try to render the SVG.
-    image_proc.render_svg(tmp_svg_path, tmp_png_path, renderer, size)
-    # try to convert to WebP.
-    image_proc.convert_webp(tmp_png_path, out_path)
 
 
+    if format == "png":
+        # single-step process
+        image_proc.render_svg(tmp_svg_path, out_path, renderer, size)
+    else:
 
-    # delete temporary files
-    os.remove(tmp_svg_path)
-    os.remove(tmp_png_path)
+        # multi-step process
+        image_proc.render_svg(tmp_svg_path, tmp_png_path, renderer, size)
 
-
-
-
-
-
-def to_avif(emoji_svg, out_path, renderer, size, name):
-    """
-    Lossless AVIF Exporting function. Creates temporary SVGs and PNGs
-    first before converting to AVIF.
-    """
-
-    tmp_svg_path = '.tmp' + name + '.svg'
-    tmp_png_path = '.tmp' + name + '.png'
-
-
-    # try to write a temporary SVG.
-    image_proc.write_temp_svg(emoji_svg, tmp_svg_path)
-    # try to render the SVG.
-    image_proc.render_svg(tmp_svg_path, tmp_png_path, renderer, size)
-    # try to convert to AVIF.
-    image_proc.convert_avif(tmp_png_path, out_path)
+        if format == "webp":
+            image_proc.convert_webp(tmp_png_path, out_path)
+        elif format == "avif":
+            image_proc.convert_avif(tmp_png_path, out_path)
+        elif format == "flif":
+            image_proc.convert_flif(tmp_png_path, out_path)
+        else:
+            os.remove(tmp_svg_path)
+            os.remove(tmp_png_path)
+            raise ValueError(f"This function wasn't given a correct format! ({format})")
 
 
     # delete temporary files
     os.remove(tmp_svg_path)
-    os.remove(tmp_png_path)
 
-
-
-
-
-
-def to_flif(emoji_svg, out_path, renderer, size, name):
-    """
-    FLIF Exporting function. Creates temporary SVGs and PNGs
-    first before converting to FLIF.
-    """
-
-    tmp_svg_path = '.tmp' + name + '.svg'
-    tmp_png_path = '.tmp' + name + '.png'
-
-    # try to write a temporary SVG.
-    image_proc.write_temp_svg(emoji_svg, tmp_svg_path)
-    # try to render the SVG.
-    image_proc.render_svg(tmp_svg_path, tmp_png_path, renderer, size)
-    # try to convert to WebP.
-    image_proc.convert_flif(tmp_png_path, out_path)
-
-
-
-    # delete temporary files
-    os.remove(tmp_svg_path)
-    os.remove(tmp_png_path)
+    if format != "png":
+        os.remove(tmp_png_path)
